is finished

+++ opinions_app/api_views.py
@@ -1,10 +1,14 @@
-# what_to_watch/opinions_app/api_views.py
+from random import randrange
 
 # Импортировать метод jsonify.
-from flask import jsonify, request
+from flask import abort, jsonify, request
 
 from . import app, db
 from .models import Opinion
+from .views import random_opinion
+
+# Импорт исключения.
+from .error_handlers import InvalidAPIUsage
 
 
 # Применять только символы из набора ASCII? Нет!
@@ -25,19 +29,26 @@ app.json.ensure_ascii = False
 # Явно разрешить метод GET.
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
-    # Получить объект по id или выбросить ошибку 404.
-    opinion = Opinion.query.get_or_404(id)
-    data = opinion.to_dict()
-    # Конвертировать данные в JSON и вернуть JSON-объект и HTTP-код ответа.
-    return jsonify({'opinion': data}), 200
+    # Получить объект по id.
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом.
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
+    return jsonify({'opinion': opinion.to_dict()}), 200
 
 
 @app.route('/api/opinions/<int:id>/', methods=['PATCH'])
 def update_opinion(id):
     data = request.get_json()
-    # Если метод get_or_404 не найдёт указанный ID,
-    # он выбросит исключение 404:
-    opinion = Opinion.query.get_or_404(id)
+    if (
+        'text' in data and
+        Opinion.query.filter_by(text=data['text']).first() is not None
+    ):
+        # Выбросить собственное исключение.
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных', 400)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     opinion.title = data.get('title', opinion.title)
     opinion.text = data.get('text', opinion.text)
     opinion.source = data.get('source', opinion.source)
@@ -46,13 +57,16 @@ def update_opinion(id):
     # Объект opinion добавлять в сессию не нужно.
     # Этот объект получен из БД методом get_or_404() и уже хранится в сессии.
     db.session.commit()
-    # При изменении объекта нужно вернуть сам объект и код 200.
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    # При изменении объекта нужно вернуть сам объект и код 201.
+    return jsonify({'opinion': opinion.to_dict()}), 201
 
 
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом.
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     # При удалении принято возвращать только код ответа 204.
@@ -70,6 +84,16 @@ def get_opinions():
 def add_opinion():
     # Получить данные из запроса в виде словаря.
     data = request.get_json()
+    if 'title' not in data or 'text' not in data:
+        # Выбросить собственное исключение.
+        # Второй параметр (статус-код) в этом обработчике можно не передавать:
+        # нужно вернуть код 400, а он и так возвращается по умолчанию.
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля', 400)
+    # Если в базе данных уже есть объект
+    # с таким же значением поля text...
+    if Opinion.query.filter_by(text=data['text']).first() is not None:
+        # Выбросить собственное исключение.
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных', 400)
     # Создать новый пустой экземпляр модели.
     opinion = Opinion()
     # Наполнить экземпляр данными из запроса.
@@ -79,3 +103,16 @@ def add_opinion():
     # Сохранить изменения.
     db.session.commit()
     return jsonify({'opinion': opinion.to_dict()}), 201
+
+
+@app.route('/api/get-random-opinion/', methods=['GET'])
+def get_random_opinion():
+    opinion = random_opinion()
+    # Если мнение найдено (переменная opinion не равна None),
+    # оно возвращается в виде JSON-объекта с кодом ответа 200 (OK).
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    # Если мнение не найдено (opinion равен None),
+    # вызывается исключение InvalidAPIUsage с сообщением об ошибке
+    # и кодом ответа 404 (Not Found).
+    raise InvalidAPIUsage('В базе данных нет мнений', 404)
